refactor: move common icon buttons to 'common' directory
