+++ dev-client/src/components/sheets/OverlaySheet.tsx
@@ -25,7 +25,7 @@ import {
 import {useFocusEffect} from '@react-navigation/native';
 
 import {BackdropComponent} from 'terraso-mobile-client/components/BackdropComponent';
-import {BigCloseButton} from 'terraso-mobile-client/components/buttons/icons/BigCloseButton';
+import {BigCloseButton} from 'terraso-mobile-client/components/buttons/icons/common/BigCloseButton';
 import {
   ModalContext,
   ModalHandle,
