+++ dev-client/src/components/tooltips/HelpTooltipButton.tsx
@@ -17,7 +17,7 @@
 
 import {memo} from 'react';
 
-import {HelpButton} from 'terraso-mobile-client/components/buttons/icons/HelpButton';
+import {HelpButton} from 'terraso-mobile-client/components/buttons/icons/common/HelpButton';
 import {Tooltip} from 'terraso-mobile-client/components/tooltips/Tooltip';
 
 type Props = React.PropsWithChildren;
